+++ TPAnualMavenizado/src/main/java/opf5/Handicap.java
@@ -2,7 +2,7 @@ package opf5;
 
 public class Handicap implements Criterio {
 
-	public int funcion(Jugador jugador) {
+	public double funcion(Jugador jugador) {
 		return jugador.handicap();
 	}
